+++ lib/meilisearch.dart
@@ -1,19 +1,10 @@
 library meilisearch;
 
+export 'src/settings/_exports.dart';
+export 'src/results/_exports.dart';
+export 'src/query_parameters/_exports.dart';
+export 'src/filter_builder/_exports.dart';
 export 'src/client.dart';
 export 'src/index.dart';
-export 'src/settings/_exports.dart';
-export 'src/facet_stat.dart';
-export 'src/match_position.dart';
-export 'src/searchable.dart';
-export 'src/multi_search_result.dart';
-export 'src/query_parameters/multi_search_query.dart';
-export 'src/query_parameters/index_search_query.dart';
-export 'src/filter_builder/_exports.dart';
-export 'src/query_parameters/_exports.dart';
-export 'src/matching_strategy_enum.dart';
-export 'src/key.dart';
 export 'src/exception.dart';
-export 'src/task.dart';
-export 'src/swap_index.dart';
 export 'src/facade.dart';
