+++ lib/src/query_parameters/indexes_query.dart
@@ -1,4 +1,4 @@
-import 'package:meilisearch/src/query_parameters/queryable.dart';
+import 'queryable.dart';
 
 class IndexesQuery extends Queryable {
   final int? offset;
