+++ lib/src/query_parameters/keys_query.dart
@@ -1,4 +1,4 @@
-import 'package:meilisearch/src/query_parameters/queryable.dart';
+import 'queryable.dart';
 
 class KeysQuery extends Queryable {
   final int? offset;
