+++ lib/src/query_parameters/multi_search_query.dart
@@ -1,4 +1,4 @@
-import 'package:meilisearch/src/query_parameters/queryable.dart';
+import 'queryable.dart';
 
 import '../annotations.dart';
 import 'index_search_query.dart';
