+++ lib/src/query_parameters/tasks_query.dart
@@ -1,4 +1,4 @@
-import 'package:meilisearch/src/query_parameters/queryable.dart';
+import 'queryable.dart';
 
 class TasksQuery extends Queryable {
   final int? from;
