+++ lib/src/results/_exports.dart
@@ -1,0 +1,14 @@
+export 'multi_search_result.dart';
+export 'searchable.dart';
+export 'tasks_results.dart';
+export 'result.dart';
+export 'facet_search_result.dart';
+export 'facet_hit.dart';
+export 'key.dart';
+export 'task.dart';
+export 'task_error.dart';
+export 'match_position.dart';
+export 'matching_strategy_enum.dart';
+export 'index_stats.dart';
+export 'all_stats.dart';
+export 'facet_stat.dart';
