+++ lib/src/results/index_stats.dart
@@ -1,0 +1,18 @@
+class IndexStats {
+  IndexStats({
+    this.numberOfDocuments,
+    this.isIndexing,
+    this.fieldsDistribution,
+  });
+
+  final int? numberOfDocuments;
+  final bool? isIndexing;
+  final Map<String, int>? fieldsDistribution;
+
+  factory IndexStats.fromMap(Map<String, Object?> map) => IndexStats(
+        numberOfDocuments: map['numberOfDocuments'] as int?,
+        isIndexing: map['isIndexing'] as bool?,
+        fieldsDistribution:
+            (map['fieldsDistribution'] as Map?)?.cast<String, int>(),
+      );
+}
