+++ lib/src/results/matching_strategy_enum.dart
