+++ lib/src/results/paginated_search_result.dart
