+++ lib/src/results/result.dart
