+++ lib/src/results/search_result.dart
