+++ lib/src/results/searchable.dart
