+++ lib/src/results/task.dart
@@ -1,3 +1,5 @@
+import 'task_error.dart';
+
 class Task {
   Task({
     this.status,
@@ -41,25 +43,4 @@ class Task {
       details: map['details'] as Map<String, Object?>?,
     );
   }
-}
-
-class TaskError {
-  TaskError({
-    this.message,
-    this.code,
-    this.type,
-    this.link,
-  });
-
-  final String? message;
-  final String? code;
-  final String? type;
-  final String? link;
-
-  factory TaskError.fromMap(Map<String, Object?> map) => TaskError(
-        message: map['message'] as String?,
-        code: map['code'] as String?,
-        type: map['type'] as String?,
-        link: map['link'] as String?,
-      );
 }
