+++ lib/src/results/tasks_results.dart
@@ -1,0 +1,30 @@
+import 'task.dart';
+
+import '../annotations.dart';
+
+class TasksResults {
+  final List<Task> results;
+  final int? next;
+  final int? limit;
+  final int? from;
+  @RequiredMeiliServerVersion('1.3.0')
+  final int? total;
+  const TasksResults({
+    this.results = const [],
+    this.limit,
+    this.from,
+    this.next,
+    this.total,
+  });
+
+  factory TasksResults.fromMap(Map<String, Object?> map) => TasksResults(
+        results: (map['results'] as Iterable)
+            .cast<Map<String, Object?>>()
+            .map((item) => Task.fromMap(item))
+            .toList(),
+        next: map['next'] as int?,
+        from: map['from'] as int?,
+        limit: map['limit'] as int?,
+        total: map['total'] as int?,
+      );
+}
