+++ test/get_client_stats_test.dart
@@ -1,4 +1,4 @@
-import 'package:meilisearch/src/task.dart';
+import 'package:meilisearch/meilisearch.dart';
 import 'package:test/test.dart';
 
 import 'utils/client.dart';
